+++ clarifai/client/app.py
@@ -86,8 +86,9 @@ class App(Lister, BaseClient):
         per_page=per_page,
         page_no=page_no)
     for dataset_info in all_datasets_info:
-      if 'version' in list(dataset_info.keys()):
-        del dataset_info['version']['metrics']
+      if 'version' in dataset_info:
+        dataset_info['version'].pop('metrics', None)
+        dataset_info['version'].pop('export_info', None)
       yield Dataset.from_auth_helper(auth=self.auth_helper, **dataset_info)
 
   def list_models(self,
@@ -457,11 +458,12 @@ class App(Lister, BaseClient):
 
     return Module.from_auth_helper(auth=self.auth_helper, module_id=module_id, **kwargs)
 
-  def dataset(self, dataset_id: str, **kwargs) -> Dataset:
+  def dataset(self, dataset_id: str, dataset_version_id: str = None, **kwargs) -> Dataset:
     """Returns a Dataset object for the existing dataset ID.
 
     Args:
         dataset_id (str): The dataset ID for the dataset to interact with.
+        dataset_version_id (str): The version ID for the dataset version to interact with.
 
     Returns:
         Dataset: A Dataset object for the existing dataset ID.
@@ -480,6 +482,7 @@ class App(Lister, BaseClient):
     kwargs = self.process_response_keys(dict_response[list(dict_response.keys())[1]],
                                         list(dict_response.keys())[1])
     kwargs['version'] = response.dataset.version if response.dataset.version else None
+    kwargs['dataset_version_id'] = dataset_version_id
     return Dataset.from_auth_helper(auth=self.auth_helper, **kwargs)
 
   def model(self, model_id: str, model_version_id: str = "", **kwargs) -> Model: