+++ clarifai/datasets/export/inputs_annotations.py
@@ -21,7 +21,7 @@ logger = get_logger("INFO", __name__)
 class DatasetExportReader:
 
   def __init__(self,
-               session: requests.Session,
+               session: requests.Session = None,
                archive_url: Optional[str] = None,
                local_archive_path: Optional[str] = None):
     """Download/Reads the zipfile archive and yields every api.Input object.
@@ -31,9 +31,11 @@ class DatasetExportReader:
         archive_url: URL of the DatasetVersionExport archive
         local_archive_path: Path to the DatasetVersionExport archive
     """
-    self.input_count = 0
+    self.input_count = None
     self.temp_file = None
     self.session = session
+    if not self.session:
+      self.session = requests.Session()
 
     assert archive_url or local_archive_path, UserError(
         "Either archive_url or local_archive_path must be provided.")
@@ -59,7 +61,8 @@ class DatasetExportReader:
   def _download_temp_archive(self, archive_url: str,
                              chunk_size: int = 128) -> tempfile.TemporaryFile:
     """Downloads the temp archive of InputBatches."""
-    r = self.session.get(archive_url, stream=True)
+    session = requests.Session()
+    r = session.get(archive_url, stream=True)
     temp_file = tempfile.TemporaryFile()
     for chunk in r.iter_content(chunk_size=chunk_size):
       temp_file.write(chunk)
@@ -67,10 +70,12 @@ class DatasetExportReader:
     return temp_file
 
   def __len__(self) -> int:
-    if not self.input_count:
+    if self.input_count is None:
+      input_count = 0
       if self.file_name_list is not None:
         for filename in self.file_name_list:
-          self.input_count += int(filename.split('_n')[-1])
+          input_count += int(filename.split('_n')[-1])
+      self.input_count = input_count
 
     return self.input_count
 
@@ -111,7 +116,8 @@ class InputAnnotationDownloader:
     """
     self.input_iterator = input_iterator
     self.num_workers = min(num_workers, 10)  # Max 10 threads
-    self.num_inputs_annotations = 0
+    self.num_inputs = 0
+    self.num_annotations = 0
     self.split_prefix = None
     self.session = session
     self.input_ext = dict(image=".png", text=".txt", audio=".mp3", video=".mp4")
@@ -182,14 +188,14 @@ class InputAnnotationDownloader:
         self._save_audio_to_archive(new_archive, hosted_url, file_name)
       elif input_type == "video":
         self._save_video_to_archive(new_archive, hosted_url, file_name)
-      self.num_inputs_annotations += 1
+      self.num_inputs += 1
 
     if data_dict.get("concepts") or data_dict.get("regions"):
       file_name = os.path.join(split, "annotations", input_.id + ".json")
       annot_data = data_dict.get("concepts") or data_dict.get("regions")
 
       self._save_annotation_to_archive(new_archive, annot_data, file_name)
-      self.num_inputs_annotations += 1
+      self.num_annotations += 1
 
   def _check_output_archive(self, save_path: str) -> None:
     try:
@@ -198,8 +204,8 @@ class InputAnnotationDownloader:
       raise e
     assert len(
         archive.namelist()
-    ) == self.num_inputs_annotations, "Archive has %d inputs+annotations | expecting %d inputs+annotations" % (
-        len(archive.namelist()), self.num_inputs_annotations)
+    ) == self.num_inputs + self.num_annotations, "Archive has %d inputs+annotations | expecting %d inputs+annotations" % (
+        len(archive.namelist()), self.num_inputs + self.num_annotations)
 
   def download_archive(self, save_path: str, split: Optional[str] = None) -> None:
     """Downloads the archive from the URL into an archive of inputs, annotations in the directory format
@@ -218,5 +224,5 @@ class InputAnnotationDownloader:
             progress.update()
 
     self._check_output_archive(save_path)
-    logger.info("Downloaded %d inputs+annotations to %s" % (self.num_inputs_annotations,
-                                                            save_path))
+    logger.info("Downloaded %d inputs and %d annotations to %s" %
+                (self.num_inputs, self.num_annotations, save_path))